requirements.txt
asgiref==3.8.1
black==24.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.9
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
flake8==6.1.0
flake8-django==1.4.0
idna==3.10
inflection==0.5.1
isort==5.13.2
mypy==1.9.0
packaging==24.2
pillow==11.0.0
psutil==6.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pylint==2.17.7
pylint-django==2.5.5
python-environ==0.4.54
pytz==2024.2
PyYAML==6.0.2
redis==5.0.1
requests==2.31.0
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
